fix: Prevent user from submitting or saving more than one application/draft for an event
When a user is not logged in and gets to the continue_as_guest? page while applying for an event, she should not be able to see the "new application" form after signing in, if there has been a previous application submission or draft storage, but instead should be redirected to their applications overview page after login.